module: inherit User and Group from infected service, enable PrivateTmp
By using PrivateTmp=true in conjunction with JoinsNamespaceOf=, it is
possible to share the private /tmp that systemd sets up. This
allows us to ~automatically get the proper permissions for reading a
secret file.

Prior to this, the key file would be dropped in
`/run/keys/files/${filename}`, which would require somehow setting up
permissions for that (e.g. give services the `keys` SupplementaryGroup,
as well as making sure all the parent directories are reachable).

One thing to take note of, is that this means any service that expects
PrivateTmp to be disabled for proper functionality will not work with
this setup. PrivateTmp is required on all services that need access to
the secret(s), and JoinsNamespaceOf should point to the sidecar service
(which currently follows the naming pattern of
`detsys-vaultAgent-${serviceName}.service`, where `${serviceName}` does
not contain the `.service` suffix).

This was initially just setting the User and Group, but that exposed
problems with nginx reading the secret file (it appears that
`/run/keys/files` was not readable to nginx or the keys group).