module: use inotifywait instead of a busy wait
We use inotifywait to wait for the secret file to be written (by
listening for the `close_write` event) under `/tmp/detsys-vault`. This
is done in the background because it's possible that there are many
files we wish to watch for.

We also use inotfywait to wait for the `/tmp/detsys-vault` directory to
exist in the first place. This is done in the foreground because the
directory must exist (i.e. assigned an inode) for inotifywait to be able
to watch it for events.

This implementation uses the shell's job control functionality to
background the non-blocking "wait for secret file" inotifywaits and then
wait for them all at once for the script to succeed. This is more
efficient than blocking on each individual secret file.