messenger: use async, allow child to return before backoff completes
Prior to this change, if the vault agent failed before backoff
completed, we would wait until the backoff completed and _then_ fail.
Since the backoff maxes out at 15 minutes, it's better to fail the unit
earlier.

Rather than dealing with threads (manually, at least), we use async-std
in order to start running the backoff in the background, allowing us to
wait on the vault agent in the foreground and retrieve its status code
if it exists with a failure.