Fix compatibility with older Nixpkgs
Prior to this change, the messenger package would not compile when
using older Nixpkgs (e.g. 21.05) because it depends on a feature only
found in newer Nixpkgs (the `cargoLock.lockFile` functionality).

By making an overlay exposing `messenger` and having the module consume
this overlay, as opposed to `pkgs.callPackage`ing the expression, we
will ensure that the proper Nixpkgs (the one we tested against and is
pinned in our flake.lock) is used.

One minor (IMHO) downside is that this does bring another Nixpkgs
instance into the system's closure.