module: invert Wants= and WantedBy= relationship
Prior to this change, the service described by the `fullServiceName`
variable would never start up if the sidecar service failed the first
time it started up, but succeeded later (we've seen this happen in AWS
EC2 instances, where Vault is not able to authenticate with AWS
immediately upon startup for whatever reason).

Now, the full service will be able to start after the sidecar service
eventually succeeds (if ever).